requirements.txt
langchain 
google.generativeai 
langchain-google-genai 
langchain-text-splitters 
requests  
langchain-experimental 
boto3 
pandas 
numpy
streamlit
aiohttp
asyncio
tabulate
scikit-learn
scipy
python-dotenv
pillow
plotly
graphviz
imblearn
seaborn
